// game-web-ui/src/app/shared/openapi/model/action.ts
/**
 * The Game - Endangered Nature Edition, built with Apodini
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Card } from './card';

export interface Action { 
    gamePileId: string;
    playerId: string;
    card: Card;
}